+++ spec/mongoid/tasks/database_spec.rb
@@ -1,9 +1,61 @@
 # frozen_string_literal: true
 
 require "spec_helper"
-require_relative './database_spec_models'
 
 describe "Mongoid::Tasks::Database" do
+
+  before(:all) do
+    module DatabaseSpec
+      class Measurement
+        include Mongoid::Document
+
+        field :timestamp, type: Time
+        field :temperature, type: Integer
+
+        embeds_many :comments
+
+        store_in collection: "measurement",
+          collection_options: {
+            capped: true, size: 10000
+          }
+      end
+
+      class Comment
+        include Mongoid::Document
+
+        field :content, type: String
+
+        embedded_in :measurement
+
+        store_in collection_options: {
+            capped: true, size: 10000
+          }
+      end
+
+      class Note
+        include Mongoid::Document
+
+        field :text
+
+        recursively_embeds_one
+
+        store_in collection_options: {
+          capped: true, size: 10000
+        }
+      end
+    end
+  end
+
+  after(:all) do
+    Mongoid.deregister_model(DatabaseSpec::Measurement)
+    DatabaseSpec.send(:remove_const, :Measurement)
+    Mongoid.deregister_model(DatabaseSpec::Comment)
+    DatabaseSpec.send(:remove_const, :Comment)
+    Mongoid.deregister_model(DatabaseSpec::Note)
+    DatabaseSpec.send(:remove_const, :Note)
+    Object.send(:remove_const, :DatabaseSpec)
+  end
+
 
   let(:logger) do
     double("logger").tap do |log|
